app/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-flask-exporter==0.23.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
